+++ CHANGELOG.rst
@@ -1,19 +1,28 @@
 Changelog for Nautilai
 ======================
 
-0.2.4 (unreleased)
+0.2.5 (unreleased)
+------------------
+
+Added:
+^^^^^^
+- GxP logging
+
+
+0.2.4 (2024-11-14)
 ------------------
 
 Added:
 ^^^^^^
 - More tooltips for invalid inputs
-- GxP logging
+- Crop AVI file to remove overlapping areas
 
 Changed:
 ^^^^^^^^
 - Prevent starting acquisition if an invalid plate ID is entered
 - Invalid inputs will now change border color instead of background color
 - Better autocomplete for Plate ID input
+- Stage is moved back to position 1 after calibration completes during start up
 
 Fixed:
 ^^^^^^